Break point
between the last commit and this has happend many many things,
i was working in the seller panel adding the product modules and many other things but in a few days ago i decided to make a drastic change in the site structure. I united the seller account with the buyer account into the user account this has made that i need to change a lot of things.
the auth service was modified now there is only a page where the users can sign up, the seller panel has a login page as also the shopping page, but the other services like password confirmation, reset password, and verify email, remaing in only one place, from the others login pages it going to redirect to those service when it is needed.
Also the lang system is changing making the lang into string keys.
Now when the users verify their email the verify email page it is redirected to the intended place.
The dark mode is been implemented in every place.

The last and very important change is that the system is going to be divided in a normal php site for the seller panel, admin panel and the home page, but the shopping page is going to be made entire in vuejs making it an api rest application